Tests/LibWeb/Text/expected/navigation/iframe-nav-before-load-creates-history-entry.txt

  step 0 iframe-nav-before-load-page-a.html (current)
  step 1 iframe-nav-before-load-page-b.html
